requirements.txt
pandas
numpy
Pillow
matplotlib
seaborn
torch
transformers
shap
captum
scikit-learn
pytest
